Add new panel component
