+++ site/components/Panel.old/index.ts
@@ -1,0 +1,2 @@
+
+export * from './Panel'
