+++ site/components/Panel.old/panel.pcss
@@ -1,0 +1,71 @@
+
+@import '../../global-css/mixins.pcss';
+
+:root {
+  --panel-default-width: 216px;
+  --panel-width: 100vw;
+  --panel-transition-time: 0.625s;
+  --panel-max-height: 100vh;
+
+  @mixin breakpoint mobile {
+    --panel-max-height: 85vh;
+  }
+}
+
+.panel {
+  color: var(--grey-90);
+  min-width: var(--panel-default-width);
+  background: var(--white);
+  position: relative;
+  flex-shrink: 1;
+
+  .swapper {
+    position: relative;
+    max-height: calc(var(--panel-max-height) - 134px - var(--panel-layout-margin));
+    transition: right 0s, height var(--panel-transition-time);
+    right: 0;
+
+    .scrollable {
+      height: 100%;
+      overflow-y: auto;
+      -ms-overflow-style: none;
+      scrollbar-width: none;
+
+      &::-webkit-scrollbar {
+        width: 0px;
+        background: transparent;
+      }
+    }
+
+    .current {
+      position: relative;
+    }
+
+    .prev {
+      position: absolute;
+      left: 0;
+      right: 0;
+      top: 0;
+    }
+
+    .next {
+      position: relative;
+      right: -100%;
+      top: 0;
+    }
+
+    &.transition {
+      right: 100%;
+      transition: right var(--panel-transition-time), height var(--panel-transition-time);
+    }
+
+    &.transition-back {
+      right: -100%;
+      transition: right var(--panel-transition-time), height var(--panel-transition-time);
+
+      .next {
+        right: 100%;
+      }
+    }
+  }
+}
